CommandLine refactoring and merge
