--- jodd-core/src/main/java/jodd/execution/NoSuchExecutorException.java
@@ -1,20 +1,0 @@
-package jodd.execution;
-
-import jodd.exception.UncheckedException;
-
-/**
- * @author Vilmos Papp
- */
-public class NoSuchExecutorException extends UncheckedException{
-	public NoSuchExecutorException(Throwable t) {
-		super(t);
-	}
-
-	public NoSuchExecutorException(String message) {
-		super(message);
-	}
-
-	public NoSuchExecutorException(String message, Throwable t) {
-		super(message, t);
-	}
-}
